requirements.txt
python-telegram-bot>=20.0
mcp>=0.1.0
gTTS>=2.3.0
python-dotenv>=1.0.0
